backend/services/resume-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
pypdf
python-docx
docx2txt
python-jose[cryptography]
passlib[bcrypt]
boto3
pika
